% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ws4sql-package.R
\docType{package}
\name{ws4sql-package}
\alias{ws4sql}
\alias{ws4sql-package}
\title{ws4sql: Read from a ws4sql API instance (which serves JSON over HTTP)}
\description{
The ws4sql server allows remote clients to issue web transactions such as queries against a slim REST API providing JSON over http(s), where the data is backed by database engines such as sqlite and duckdb. This R package provides a function to issue queries to such remote API instances and get results back in tabular form.
}
\author{
\strong{Maintainer}: Markus Skyttner \email{markussk@kth.se}

}
\keyword{internal}
