% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-client.R
\name{ws4sql_con}
\alias{ws4sql_con}
\title{Create a connection to a ws4sql server instance}
\usage{
ws4sql_con(connectionstring = NULL, quiet = TRUE, api_key)
}
\arguments{
\item{connectionstring}{the connectionstring to use for the ws4sql server}

\item{quiet}{a boolean indicating whether some messages should be displayed}

\item{api_key}{a character only provided if using an API key}
}
\value{
a connection object which can be used with the other functions
}
\description{
Create a connection to a ws4sql server instance
}
\details{
The connectionstring should follow this pattern:
http(s)://user:pass@hostname/path/

Examples:

http://localhost/duckserve/
https://user:pass@some.server.tld:8000/duckserve/

When not using basic auth (like above), use the api_key parameter
}
