% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-client.R
\name{ws4sql_ping}
\alias{ws4sql_ping}
\title{Check that a ws4sql instance is available}
\usage{
ws4sql_ping(con = ws4sql_con(quiet = TRUE))
}
\arguments{
\item{con}{the connection for the ws4sql, see ws4sql_con()}
}
\value{
TRUE if available, otherwise FALSE
}
\description{
Check that a ws4sql instance is available
}
