% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-client.R
\name{ws4sql_read}
\alias{ws4sql_read}
\title{Read data from ws4sql}
\usage{
ws4sql_read(con = ws4sql_con(quiet = TRUE), statement = NULL)
}
\arguments{
\item{con}{the connection, by default from duckhttp_con()}

\item{statement}{the web transaction to issue against the ws4sql httpserver}
}
\value{
a tibble with results from the query
}
\description{
Read data from ws4sql
}
